# plugins/llm-service/openapi.yaml
openapi: 3.1.0
info:
  title: LLM Service API
  description: |
    A production-ready FastAPI wrapper around vLLM for on-premise LLM inference.
    
    ## Features
    - Text generation with streaming
    - Chat completion
    - Text classification
    - Entity extraction
    - Text summarization
    - Embeddings generation
    - **Vision OCR** - Intelligent text extraction from images
    - **Data Transform** - LLM-powered data transformation
    
    ## Authentication
    API key authentication via `X-API-Key` header (when enabled).
    
  version: 1.1.0
  contact:
    name: FlowForge Team
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.flowforge.io/llm
    description: Production server
...
paths:
...
  /api/v1/transform:
    post:
      tags: [Transform]
      summary: Transform data
      description: |
        Transform data using natural language instructions.
        The LLM understands your intent and applies the transformation intelligently.
      operationId: transformData
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransformRequest'
      responses:
        '200':
          description: Successful transformation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransformResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/transform/schema:
    post:
      tags: [Transform]
      summary: Transform to target schema
      description: Transform data to match a target JSON schema with intelligent field mapping.
      operationId: transformToSchema
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchemaTransformRequest'
      responses:
        '200':
          description: Successful schema transformation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaTransformResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/transform/convert:
    post:
      tags: [Transform]
      summary: Convert data format
      description: Convert data between formats (JSON, XML, YAML, CSV, etc.)
      operationId: convertFormat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FormatConvertRequest'
      responses:
        '200':
          description: Successful conversion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FormatConvertResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/transform/clean:
    post:
      tags: [Transform]
      summary: Clean and normalize data
      description: Clean and normalize data with operations like trim, deduplicate, standardize dates, etc.
      operationId: cleanData
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CleanDataRequest'
      responses:
        '200':
          description: Successful cleaning
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CleanDataResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/transform/merge:
    post:
      tags: [Transform]
      summary: Merge multiple data sources
      description: Intelligently merge multiple data sources with conflict resolution.
      operationId: mergeData
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MergeDataRequest'
      responses:
        '200':
          description: Successful merge
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MergeDataResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          default: false
        error:
          type: string
        code:
          type: string
        details:
          type: object

    # Vision Schemas
...
    TransformRequest:
      type: object
      required:
        - data
        - instruction
      properties:
        data: {}
        instruction:
          type: string
          minLength: 5
        input_format:
          type: string
          enum: [json, xml, yaml, csv, markdown, text, html, sql]
        output_format:
          type: string
          enum: [json, xml, yaml, csv, markdown, text, html, sql]
          default: json
        examples:
          type: array
          items:
            type: object
            properties:
              input: {}
              output: {}
        model:
          type: string
    TransformResponse:
      type: object
      properties:
        success:
          type: boolean
        result: {}
        input_format:
          type: string
        output_format:
          type: string
        model:
          type: string
    SchemaTransformRequest:
      type: object
      required:
        - data
        - target_schema
      properties:
        data: {}
        source_schema:
          type: object
        target_schema:
          type: object
        mapping_hints:
          type: object
          additionalProperties:
            type: string
        fill_defaults:
          type: boolean
          default: true
        model:
          type: string
    SchemaTransformResponse:
      type: object
      properties:
        success:
          type: boolean
        result: {}
        mappings_used:
          type: object
        fields_filled:
          type: array
          items:
            type: string
        validation_passed:
          type: boolean
        model:
          type: string
    FormatConvertRequest:
      type: object
      required:
        - data
        - target_format
      properties:
        data:
          type: string
        source_format:
          type: string
          enum: [json, xml, yaml, csv, markdown, text, html, sql]
        target_format:
          type: string
          enum: [json, xml, yaml, csv, markdown, text, html, sql]
        infer_structure:
          type: boolean
          default: true
        model:
          type: string
    FormatConvertResponse:
      type: object
      properties:
        success:
          type: boolean
        result:
          type: string
        source_format:
          type: string
        target_format:
          type: string
        inferred_schema:
          type: object
        model:
          type: string
    CleanDataRequest:
      type: object
      required:
        - data
        - operations
      properties:
        data: {}
        operations:
          type: array
          items:
            type: string
            enum:
              - trim
              - normalize_whitespace
              - fix_encoding
              - standardize_dates
              - standardize_phones
              - standardize_addresses
              - deduplicate
              - fill_missing
              - fix_typos
              - normalize_case
          minItems: 1
        locale:
          type: string
          default: en-US
        model:
          type: string
    CleanDataResponse:
      type: object
      properties:
        success:
          type: boolean
        result: {}
        changes_made:
          type: array
          items:
            type: string
        records_removed:
          type: integer
        model:
          type: string
    MergeDataRequest:
      type: object
      required:
        - sources
      properties:
        sources:
          type: array
          items: {}
          minItems: 2
        merge_strategy:
          type: string
          enum: [union, intersection, left, smart]
          default: smart
        key_fields:
          type: array
          items:
            type: string
        conflict_resolution:
          type: string
          enum: [first, last, merge, ask]
          default: merge
        model:
          type: string
    MergeDataResponse:
      type: object
      properties:
        success:
          type: boolean
        result: {}
        sources_merged:
          type: integer
        conflicts_resolved:
          type: integer
        merge_report:
          type: string
        model:
          type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key